Sample edge list format for this program
This is an edge list for the DIrected network included by S. Hansell in "Cooperative groups, weak ties, and the integration of peer friendships" (1984). The input code for this program is not very sophisticated so users should change their network formatting to match the following requirements when inputting:

Vertex numbering must start at zero and no numbers can be skipped. For example, if there are N vertices, they should be numbered {0,1,2,....,(N-1)} with no missing integers in the set. If vertex 0<=L<=(N-1) is missing from the edge list, the program may accept the data but it will load the edges to incorrect positions and produce corrupted results.

This code is for directed networks, so each edge should appear only once in the edge list. If an edge extends from vertex 0 to vertex 1, then there will be one line on the edge list: 0 1. If an edge also extends from vertex 1 to vertex 0, then there will be another line in the edge list: 1 0. Otherwise, this line should not exist. 

The edge list should be sorted least-to-greatest first on the left column and then likewise on the right column.